Move types into their own file.
