+++ protoc-gen-swagger/genswagger/template.go
@@ -11,91 +11,6 @@ import (
 	"github.com/gengo/grpc-gateway/protoc-gen-grpc-gateway/descriptor"
 )
 
-type param struct {
-	*descriptor.File
-	reg *descriptor.Registry
-}
-
-type binding struct {
-	*descriptor.Binding
-}
-
-type swaggerInfoObject struct {
-	Version string `json:"version"`
-	Title   string `json:"title"`
-}
-
-type swaggerObject struct {
-	Swagger     string                   `json:"swagger"`
-	Info        swaggerInfoObject        `json:"info"`
-	Host        string                   `json:"host,omitempty"`
-	BasePath    string                   `json:"basePath,omitempty"`
-	Schemes     []string                 `json:"schemes"`
-	Consumes    []string                 `json:"consumes"`
-	Produces    []string                 `json:"produces"`
-	Paths       swaggerPathsObject       `json:"paths"`
-	Definitions swaggerDefinitionsObject `json:"definitions"`
-}
-
-type swaggerPathsObject map[string]swaggerPathItemObject
-type swaggerPathItemObject struct {
-	Get    *swaggerOperationObject `json:"get,omitempty"`
-	Delete *swaggerOperationObject `json:"delete,omitempty"`
-	Post   *swaggerOperationObject `json:"post,omitempty"`
-	Put    *swaggerOperationObject `json:"put,omitempty"`
-}
-
-type swaggerOperationObject struct {
-	Summary    string                  `json:"summary"`
-	Responses  swaggerResponsesObject  `json:"responses"`
-	Parameters swaggerParametersObject `json:"parameters,omitempty"`
-}
-
-type swaggerParametersObject []swaggerParameterObject
-type swaggerParameterObject struct {
-	Name        string `json:"name"`
-	Description string `json:"description,omitempty"`
-	In          string `json:"in,omitempty"`
-	Required    bool   `json:"required"`
-	Type        string `json:"type,omitempty"`
-	Format      string `json:"format,omitempty"`
-
-	// Or you can explicitly refer to another type. If this is defined all
-	// other fields should be empty
-	Schema *swaggerSchemaObject `json:"schema,omitempty"`
-}
-
-type swaggerResponsesObject map[string]swaggerResponseObject
-
-type swaggerResponseObject struct {
-	Description string              `json:"description"`
-	Schema      swaggerSchemaObject `json:"schema"`
-}
-
-type swaggerSchemaObject struct {
-	Ref        string                                 `json:"$ref,omitempty"`
-	Type       string                                 `json:"type,omitempty"`
-	Properties map[string]swaggerSchemaPropertyObject `json:"properties,omitempty"`
-}
-
-type swaggerSchemaPropertyObject struct {
-	//Name   string `json:"name"`
-	Type   string `json:"type,omitempty"`
-	Format string `json:"format,omitempty"`
-
-	// Or you can explicitly refer to another type. If this is defined all
-	// other fields should be empty
-	Ref string `json:"$ref,omitempty"`
-}
-
-type swaggerReferenceObject struct {
-	Ref string `json:"$ref"`
-}
-
-type swaggerDefinitionsObject map[string]swaggerSchemaObject
-
-type messageMap map[string]*descriptor.Message
-
 // Discover all possible Messages that can be generated by the services.
 func findServicesMessages(s []*descriptor.Service, reg *descriptor.Registry, m messageMap) {
 	for _, svc := range s {
@@ -126,13 +41,6 @@ func findNestedMessages(message *descriptor.Message, reg *descriptor.Registry, m
 			}
 		}
 	}
-
-	//for _, svc := range message.NestedType {
-	//	for _, meth := range svc.Methods {
-	//		toReturn[meth.RequestType.FQMN()] = meth.RequestType
-	//		toReturn[meth.ResponseType.FQMN()] = meth.RequestType
-	//	}
-	//}
 }
 
 func renderMessagesAsDefinition(messages messageMap) swaggerDefinitionsObject {
@@ -380,15 +288,6 @@ func applyTemplate(p param) (string, error) {
 	m := messageMap{}
 	findServicesMessages(p.Services, p.reg, m)
 	s.Definitions = renderMessagesAsDefinition(m)
-
-	//s.Definitions[leak] = swaggerSchemaObject{
-	//	Properties: map[string]swaggerSchemaPropertyObject{
-	//		"hardcoded": swaggerSchemaPropertyObject{
-	//			Type:   "string",
-	//			Format: "string",
-	//		},
-	//	},
-	//}
 
 	w := bytes.NewBuffer(nil)
 	enc := json.NewEncoder(w)
